reorganize helpers in preparation for documenting them
