+++ c/driver/sqlite/sqlite.cc
@@ -27,7 +27,6 @@
 #include "driver/common/options.h"
 #include "driver/common/utils.h"
 #include "driver/framework/base_driver.h"
-#include "driver/framework/catalog.h"
 #include "driver/framework/connection.h"
 #include "driver/framework/database.h"
 #include "driver/framework/statement.h"
